Tailor the project structure to suit my needs
