+++ app/build.gradle.kts
@@ -34,7 +34,7 @@ java {
 
 application {
     // Define the main class for the application.
-    mainClass = "org.example.App"
+    mainClass = "dev.madpang.App"
 }
 
 tasks.named<Test>("test") {
