+++ app/src/main/java/dev/madpang/App.java
@@ -1,0 +1,21 @@
+/**
+ * @file: App.java
+ * @brief: Main application class for the MMD to HTML parser.
+ * @author: madpang
+ * @date:
+ * - created on 2025-06-09
+ * - updated on 2025-06-09
+ */
+
+package dev.madpang;
+
+public class App {
+	public String printInfo() {
+		return "MMD to HTML parser started.";
+	}
+
+	public static void main(String[] args) {
+		System.out.println(new App().printInfo());
+		/// @todo: Entry point for parser (to be implemented)
+	}
+}
