--- app/src/main/java/org/example/App.java
@@ -1,14 +1,0 @@
-/*
- * This source file was generated by the Gradle 'init' task
- */
-package org.example;
-
-public class App {
-    public String getGreeting() {
-        return "Hello World!";
-    }
-
-    public static void main(String[] args) {
-        System.out.println(new App().getGreeting());
-    }
-}
