+++ app/src/test/java/dev/madpang/AppTest.java
@@ -1,0 +1,20 @@
+/**
+ * @file: AppTest.java
+ * @brief: Unit tests for the App class.
+ * @author: madpang
+ * @date:
+ * - created on 2025-06-09
+ * - updated on 2025-06-09
+ */
+
+package dev.madpang;
+
+import org.junit.jupiter.api.Test;
+import static org.junit.jupiter.api.Assertions.*;
+
+class AppTest {
+    @Test void appHasAGreeting() {
+        App classUnderTest = new App();
+        assertNotNull(classUnderTest.printInfo(), "app should have a greeting");
+    }
+}
